requirements.txt
PyYAML
pydantic
paramiko
requests
tenacity
rich